reshuffle code to work with new structure
