+++ .gitignore
@@ -8,7 +8,7 @@
 .env
 /build
 /dist
-/test-reports
+/.reports
 __pycache__
 _resource
 admin_templates
